Adds interfaces for serialization of user classes
This branch adds `JsonSerialization` and `JsonTreeSerialization` interfaces.
The former provides a mechanism for instances to serialize themselves to a
`JsonWriter`, and the latter allows for serialization to a `JsonElement`.
Classes may implement one or both of the interfaces; the `JsonWriter`
version gets precedence.

This change takes effect only when the Gson instance has been built with a
GsonBuilder on which the enableSerializationInterfaces method has been
called.

There are two main benefits to this approach:

* A user class can now serialize itself without the need for a second
  auxiliary class (i.e., a TypeAdapter or a JsonSerializer).
* In conjunction with the json-constructor feature, all of the logic for
  marshaling objects to and from JSON can now live in the object itself.